// js/stat-evaluation-player/src/generated/KickoffPlayerStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LabeledCounts } from "./LabeledCounts.ts";

export type KickoffPlayerStats = { count: number, touches: number, fakes: number, misses: number, support_go_for_boosts: number, support_cheats: number, support_other: number, kickoff_goal_count: number, boost_after_sample_count: number, cumulative_boost_after: number, labeled_event_counts?: LabeledCounts, };